Simplified validation of experiments file format
